Extract reusable methods for form validation
