+++ src/components/LoginForm.jsx
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
 import Input from './common/Input';
 import Joi from 'joi-browser';
+import Form from './common/Form';
 
-class LoginForm extends Component {
+class LoginForm extends Form {
   state = {
-    account: {username: '', password: '',},
+    data: {username: '', password: '',},
     errors: {},
   }
 
@@ -13,53 +14,13 @@ class LoginForm extends Component {
     password: Joi.string().required().label('Password'),
   }
 
-  validate = () => {
-    const options = { abortEarly: false };
-    const { error } = Joi.validate(this.state.account, this.schema, options);
-
-    console.log(error);
-
-    if (!error) return null;
-
-    const errors = {};
-    for (let item of error.details)
-      errors[item.path[0]] = item.message;
-      return errors;
-  }
-
-  handleSubmit = e => {
-    e.preventDefault();
-
-    const errors = this.validate();
-    this.setState({ errors: errors || {} });
-    if (errors) return;
-
-
+  doSubmit = () => {
     // call server
     console.log("Submitted");
   }
 
-  validateProperty = ({name, value}) => {
-    const obj = { [name]: value};
-    const schema = { [name]: this.schema[name] }
-    const { error } = Joi.validate(obj, schema);
-    return error ? error.details[0].message : null;
-  }
-
-  handleChange = ({ currentTarget: input }) => {
-    const errors = { ...this.state.errors }
-    const errorMessage = this.validateProperty(input);
-
-    if (errorMessage) errors[input.name] = errorMessage;
-    else delete errors[input.name];
-
-    const account = {...this.state.account};
-    account[input.name] = input.value;
-    this.setState({ account, errors });
-  }
-
   render() { 
-    const { account, errors } = this.state;
+    const { data, errors } = this.state;
 
     return ( 
     <div>
@@ -67,14 +28,14 @@ class LoginForm extends Component {
       <form onSubmit={this.handleSubmit}>
         <Input 
           name="username" 
-          value={account.username} 
+          value={data.username} 
           label="Username" 
           error={errors.username}
           onChange={this.handleChange} 
         />
         <Input 
           name="password" 
-          value={account.password} 
+          value={data.password} 
           label="Password" 
           error={errors.password}
           onChange={this.handleChange}
